worker/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
IMDbPY @ git+https://github.com/alberanid/imdbpy@58e9781afedcefe37865b614babb271fb55d8a70
lxml==4.6.2
pycparser==2.20
pyOpenSSL==20.0.1
requests==2.25.1
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.3.22
urllib3==1.26.2
wincertstore==0.2
